4th axis handling based on the latest available grbl to date v1.1f: https://github.com/gnea/grbl
All credits to Bob Beattie for the original work on an older version of GRBL.
This is no more than re-applying his work onto the lastest version.

WARNING! There is a difference though. On the board we are using, the 4th axis is on a different port than XYZ.
This is not supported by GRBL by default.
To make it work for this particular board in this particular case, 5 "Dirty hacks" have been fitted in the code.
They can easily be found by searching for "Dirty" in stepper.c.

If there is demand, I will consider cleaning that up, and submitting a pull request to make OpenPnP-Grbl up-to-date.